Bert_TextCNN_CLS/requirements.txt
pytorch~=1.12.1
cudatoolkit~=11.3.1
tqdm~=4.64.1
transformers~=4.24.0
matplotlib~=3.5.1
scikit-learn~=1.1.1